Bugfixes for v2.0.0-beta - Correctly parses and uses the specified chromosomes - Correctly uses minimum value for shortest tract-length bin. - Correctly generates start parameters for fixed-ancestry proportion models.
