+++ example/2pops/ASW_fix_one_pulse.yaml
@@ -12,11 +12,10 @@ samples:
 model_filename: pp.yaml
 start_params: 
   tx: 6-8
-repetitions: 3
+repetitions: 2
 seed: 100
 exclude_tracts_below_cM: 2
 time_scaling_factor: 100 #Times will be scaled down by this much when running the optimizer.
-fix_ancestry_proportions: True
-params_to_fix: ['R'] 
+fix_parameters_from_ancestry_proportions: ['R'] 
 output_directory: .\output_fix\
 output_filename_format: "one_pulse_{label}"
