+++ tracts/__init__.py
@@ -1,6 +1,6 @@
 from tracts.core import *
 from tracts.parametrized_demography import ParametrizedDemography, FixedAncestryDemography
-from tracts import legacy_models
-from tracts import driver
+from . import legacy_models
+from . import driver
 #from tracts.driver import run_tracts
 run_tracts = driver.run_tracts
