+++ tracts/driver.py
@@ -11,10 +11,10 @@ import numbers
 logger = logging.getLogger(__name__)
 
 def locate_file_path(filename, absolute_driver_yaml_path=None):
-    for filepath, method_name in ((Path(filename), 'from working directory.'), 
-                                  (Path(__main__.__file__).parent / filename, 'using script directory.'), 
+    for filepath, method_name in ((Path(filename), 'using working directory'), 
+                                  (Path(__main__.__file__).parent / filename, 'using script directory'), 
                                   (Path(sys.path[0]) / filename, 'using sys.path[0]'),
-                                  (absolute_driver_yaml_path.parent / filename if isinstance(absolute_driver_yaml_path, Path) else Path(''), 'using driver yaml.')
+                                  (absolute_driver_yaml_path.parent / filename if isinstance(absolute_driver_yaml_path, Path) else Path(''), 'using driver yaml')
                                 ):
         #logger.info(f'{method_name}: {filepath}')
         if filepath.is_file():
@@ -24,47 +24,39 @@ def locate_file_path(filename, absolute_driver_yaml_path=None):
 
 def run_tracts(driver_filename):
     driver_path = locate_file_path(driver_filename)
-    if driver_path is None:
-        raise OSError('Driver yaml file could not be found.')
-    driver_spec = None
-    with driver_path.open() as file, ruamel.yaml.YAML(typ="safe") as yaml:
-        driver_spec = yaml.load(file)
-    assert isinstance(driver_spec, dict), ".yaml file was invalid."
-    directory = driver_spec['samples']['directory']
-    names = driver_spec['samples']['individual_names']
+    driver_spec = load_driver_file(driver_path)
+
     chromosome_list = parse_chromosomes(driver_spec['samples']['chromosomes'])
+    logger.info(f'Chromosomes: {chromosome_list}')
     individual_filenames = parse_individual_filenames(driver_spec['samples']['individual_names'], driver_spec['samples']['filename_format'], labels = driver_spec['samples']['labels'], directory=driver_spec['samples']['directory'], absolute_driver_yaml_path=driver_path)
+
+    exclude_tracts_below_cM = driver_spec['exclude_tracts_below_cm'] if 'exclude_tracts_below_cm' in driver_spec else 10
 
     # load the population
     pop = tracts.population(filenames_by_individual=individual_filenames, selectchrom=chromosome_list)
-    (bins, data) = pop.get_global_tractlengths(npts=50)
+    (bins, data) = pop.get_global_tractlengths(npts=50, exclude_tracts_below_cM=exclude_tracts_below_cM)
 
     logger.info(f'Bins: {bins}')
-    Ls = pop.Ls
-    nind = pop.nind
 
     time_scaling_factor = driver_spec['time_scaling_factor'] if 'time_scaling_factor' in driver_spec else 1
 
-    if 'model_filename' not in driver_spec:
-        raise ValueError('You must specify the file path to your model under "model_filename".')
-    model_path = locate_file_path(driver_spec['model_filename'], driver_path)
-    if model_path is None:
-        raise OSError('Model yaml file could not be found.')
-    model = tracts.ParametrizedDemography.load_from_YAML(model_path.resolve())
+    model = load_model_from_driver(driver_spec, driver_path)
+
     func = lambda params: model.get_migration_matrix(scale_select_indices(params, model.is_time_param(), time_scaling_factor))
     bound = lambda params: model.out_of_bounds(scale_select_indices(params, model.is_time_param(), time_scaling_factor))
 
     population_labels = model.population_indices.keys()
 
-    if 'fix_ancestry_proportions' in driver_spec and driver_spec['fix_ancestry_proportions'] == True:
+    logger.info(f'Model Parameters: {model.free_param_names}')
+
+    if 'fix_parameters_from_ancestry_proportions' in driver_spec:
         ancestry_proportions = calculate_ancestry_proportions(pop, population_labels)
-        if 'params_to_fix' not in driver_spec:
-            raise KeyError('You must specify which parameters to calculate from known ancestry proportions under "params_to_fix"')
-        model.fix_ancestry_proportions(driver_spec['params_to_fix'], ancestry_proportions)
+        model.fix_ancestry_proportions(driver_spec['fix_parameters_from_ancestry_proportions'], ancestry_proportions)
 
     if type(driver_spec['start_params']) is not dict:
-        raise KeyError('You must specify initial parameters or parameter ranges under "start_params".')    
-    params_found, likelihoods = run_model_multi_params(func, bound, pop, population_labels, parse_start_params(driver_spec['start_params'], driver_spec['repetitions'], driver_spec['seed'], model, time_scaling_factor))
+        raise KeyError('You must specify initial parameters or parameter ranges under "start_params".')
+       
+    params_found, likelihoods = run_model_multi_params(func, bound, pop, population_labels, parse_start_params(driver_spec['start_params'], driver_spec['repetitions'], driver_spec['seed'], model, time_scaling_factor), exclude_tracts_below_cM=exclude_tracts_below_cM)
     print("likelihoods found: ", likelihoods)
     optimal_params = min(zip(params_found,likelihoods), key=lambda x: x[1])[0]
     optimal_params = scale_select_indices(optimal_params, model.is_time_param(), time_scaling_factor)
@@ -72,16 +64,41 @@ def run_tracts(driver_filename):
         output_simulation_data(pop, optimal_params, model, driver_spec)
     return
 
-def parse_chromosomes(chromosome_spec):
+def load_driver_file(driver_path):
+    if driver_path is None:
+        raise OSError('Driver yaml file could not be found.')
+    driver_spec = None
+    with driver_path.open() as file, ruamel.yaml.YAML(typ="safe") as yaml:
+        driver_spec = yaml.load(file)
+    if not isinstance(driver_spec, dict):
+        raise ValueError('Driver yaml file was invalid.')
+    return driver_spec
+
+def load_model_from_driver(driver_spec, driver_path):
+    if 'model_filename' not in driver_spec:
+        raise ValueError('You must specify the file path to your model under "model_filename".')
+    model_path = locate_file_path(driver_spec['model_filename'], driver_path)
+    if model_path is None:
+        raise OSError('Model yaml file could not be found.')
+    model = tracts.ParametrizedDemography.load_from_YAML(model_path.resolve())
+    return model
+    
+def parse_chromosomes(chromosome_spec, chromosomes=None):
+    if chromosomes == None:
+        chromosomes = []
     if isinstance(chromosome_spec, int):
-        return str(chromosome_spec)
+        chromosomes.append(chromosome_spec)
+        return chromosomes
     if isinstance(chromosome_spec, list):
-        return [parse_chromosomes(subspec) for subspec in chromosome_spec]
+        [parse_chromosomes(subspec, chromosomes) for subspec in chromosome_spec]
+        return chromosomes
     try:
         chromosome_spec = chromosome_spec.split('-')
-        return
+        chromosomes.extend(range(int(chromosome_spec[0]), int(chromosome_spec[1])+1))
+        return chromosomes
     except Exception as e:
         raise ValueError('Chromosomes should be an int, range (ie: 1-22), or list.') from e
+    
     
 def parse_individual_filenames(individual_names, filename_string, labels = ['A', 'B'], directory='', absolute_driver_yaml_path=None):
     def _find_individual_file(individual_name, label_val):
@@ -93,11 +110,15 @@ def parse_individual_filenames(individual_names, filename_string, labels = ['A',
     return individual_filenames
 
 def parse_start_params(start_param_bounds, repetitions=1, seed=None, model: tracts.ParametrizedDemography=None, time_scaling_factor=1):
-    num_params = len(model.free_param_names) - len(model.params_fixed_by_ancestry)
+    #num_params = len(model.free_param_names) - len(model.params_fixed_by_ancestry)
+    num_params = len(model.free_param_names)
     rng = numpy.random.default_rng(seed=seed)
     start_params = rng.random((repetitions, num_params))
     for param_name, param_info in model.free_param_names.items():
-        if param_name not in start_param_bounds and param_name not in model.params_fixed_by_ancestry:
+        if param_name in model.params_fixed_by_ancestry:
+            start_params[:,param_info['index']] = 0
+            continue
+        if param_name not in start_param_bounds:
             raise KeyError(f"Initial values were not specified for parameter '{param_name}'.")
         if isinstance(start_param_bounds[param_name], numbers.Number):
             start_params[:,param_info['index']] = start_param_bounds[param_name]
@@ -112,15 +133,15 @@ def parse_start_params(start_param_bounds, repetitions=1, seed=None, model: trac
         if param_info['type'] == 'time':
             start_params[:,param_info['index']] *= 1/time_scaling_factor
     logger.info(f'Start Params: {start_params}')
+    if model.params_fixed_by_ancestry is not None:
+        start_params = numpy.transpose([start_params[:,param_info['index']] for param_name, param_info in model.free_param_names.items() if param_name not in model.params_fixed_by_ancestry])
+    logger.info(f'Start Params: {start_params}')
     return start_params
-            
-        
-        
-    
-        
+
         
 def scale_select_indices(arr, indices_to_scale, scaling_factor = 1):
-    assert len(indices_to_scale) == len(arr)
+    if len(indices_to_scale) != len(arr):
+        raise ValueError(f'Length of array ({len(arr)}) was not equal to length of indices_to_scale ({len(indices_to_scale)}).')
     return (numpy.multiply(indices_to_scale, scaling_factor - 1) + 1) * arr
 
 def randomize(arr, a, b):
@@ -128,21 +149,21 @@ def randomize(arr, a, b):
     # uniformly.
     return ((b-a) * numpy.random.random(arr.shape) + a)*arr
 
-def run_model_multi_params(model_func, bound_func, population, population_labels, startparams_list, cutoff = 2):
+def run_model_multi_params(model_func, bound_func, population, population_labels, startparams_list, exclude_tracts_below_cM=0):
     optimal_params = []
     likelihoods = []
     for start_params in startparams_list:
         logger.info(f'Start params: {start_params}')
-        params_found, likelihood_found = run_model(model_func, bound_func, population, population_labels, start_params)
+        params_found, likelihood_found = run_model(model_func, bound_func, population, population_labels, start_params, exclude_tracts_below_cM=exclude_tracts_below_cM)
         optimal_params.append(params_found)
         likelihoods.append(likelihood_found)
     return optimal_params, likelihoods
 
-def run_model(model_func, bound_func, population, population_labels, startparams):
+def run_model(model_func, bound_func, population, population_labels, startparams, exclude_tracts_below_cM=0):
     cutoff = 2
     Ls = population.Ls
     nind = population.nind
-    (bins, data) = population.get_global_tractlengths(npts=50)
+    (bins, data) = population.get_global_tractlengths(npts=50, exclude_tracts_below_cM=exclude_tracts_below_cM)
     data = [data[poplab] for poplab in population_labels]
     xopt = tracts.optimize_cob(startparams, bins, Ls, data, nind, model_func, outofbounds_fun=bound_func, cutoff=cutoff, epsilon=1e-2)
     optmod = tracts.demographic_model(model_func(xopt))
@@ -161,7 +182,8 @@ def output_simulation_data(sample_population, optimal_params, model: tracts.Para
         output_dir = ''
 
     output_filename_format = driver_spec['output_filename_format']
-    (bins, data) = sample_population.get_global_tractlengths(npts=50)
+    exclude_tracts_below_cM = driver_spec['exclude_tracts_below_cm'] if 'exclude_tracts_below_cm' in driver_spec else 10
+    (bins, data) = sample_population.get_global_tractlengths(npts=50, exclude_tracts_below_cM=exclude_tracts_below_cM)
     Ls = sample_population.Ls
     nind = sample_population.nind
