+++ tracts/legacy_models/__init__.py
@@ -1,1 +1,4 @@
 from . import models_2pop
+from . import models_3pop
+from . import models_4pop
+from . import models_notvalidated
